Split proxy-init into separate repo
Fixes #2563

The new repo is https://github.com/linkerd/linkerd2-proxy-init, and I
tagged the latest there `stable-2.3.0`. The plan would be to tag it each
time it is updated, using the same tag of the linkerd2 repo that
required that change. I think that helps keeping track of which
proxy-init version is required by which linkerd2 version. Let me know if
you have a different opinion!

Here, I've removed the `/proxy-init` dir and pinned the injected
proxy-init version to `stable-2.3.0` in the injector code and tests.

`/cni-plugin` depends on proxy-init, so I updated the import paths
there, and could verify CNI is still working (there is some flakiness
but unrelated to this PR).

For consistency, I added a `--init-image-version` flag to `linkerd
inject` along with its corresponding override config annotation. If
deemed too much, we can hide it (along with `--init-image`).

Signed-off-by: Alejandro Pedraza <alejandro@buoyant.io>